Fix the command line usage - Add 'lxml' as parser for bs4 - Separate scrappers in more filtered format - Use absolute paths for imports
