+++ proxygrab/cmdline/__init__.py
@@ -1,0 +1,5 @@
+from proxygrab.cmdline.core import clicmd
+
+
+def __useless_func():
+    clicmd()
